refactor: rename runtime app layer
