+++ runtime-kotlin/src/main/kotlin/skillbill/di/RuntimeComponent.kt
@@ -3,10 +3,10 @@ package skillbill.di
 import me.tatarka.inject.annotations.Component
 import me.tatarka.inject.annotations.Provides
 import skillbill.RuntimeContext
-import skillbill.app.LearningService
-import skillbill.app.ReviewService
-import skillbill.app.SystemService
-import skillbill.app.TelemetryService
+import skillbill.application.LearningService
+import skillbill.application.ReviewService
+import skillbill.application.SystemService
+import skillbill.application.TelemetryService
 
 @Component
 abstract class RuntimeComponent(
